[feat/refactor] 간병인 프로필 보내주는 정보 수정
- 임금 정보 추가
- 기본정보 제외 프로필 정보만 보내주기로 수정
- url변경 profile -> resume
- id 타입 변경 int -> Long

+++ src/main/java/com/suitecare/suitecare/api/mate/mapper/CareerMapper.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 @Mapper
 public interface CareerMapper {
-    List<CareerDTO> findCareerById(int id);
+    List<CareerDTO> findCareerById(Long id);
 }
